refactor(domain): update test imports from foundation to domain modules
Update all test-file imports to reference the new domain module paths
for ValueObject, Messages, and Specification classes.
