+++ tests/fixtures/simple_fake_command_with_value.py
@@ -1,7 +1,7 @@
 from typing import Self
 
-from forging_blocks.foundation.messages.command import Command
-from forging_blocks.foundation.messages.message import MessageMetadata
+from forging_blocks.domain.messages.command import Command
+from forging_blocks.domain.messages.message import MessageMetadata
 
 
 class SimpleFakeCommandWithValue(Command[dict[str, object]]):
