+++ tests/forging_blocks/application/ports/outbound/test_command_sender.py
@@ -13,8 +13,8 @@ from typing import Self
 import pytest
 
 from forging_blocks.application import CommandSenderPort
+from forging_blocks.domain.messages import Command, MessageMetadata
 from forging_blocks.foundation import OutboundPort
-from forging_blocks.foundation.messages import Command, MessageMetadata
 
 
 class FakeCommand(Command[str]):
