+++ tests/forging_blocks/infrastructure/message_bus/test_in_memory_message_bus.py
@@ -2,9 +2,9 @@ from typing import Any, Self
 
 import pytest
 
-from forging_blocks.foundation.messages.command import Command
-from forging_blocks.foundation.messages.message import MessageMetadata
-from forging_blocks.foundation.messages.query import Query
+from forging_blocks.domain.messages.command import Command
+from forging_blocks.domain.messages.message import MessageMetadata
+from forging_blocks.domain.messages.query import Query
 from forging_blocks.infrastructure.message_bus.in_memory_message_bus import (
     InMemoryMessageBus,
 )
