+++ tests/forging_blocks/infrastructure/message_bus/test_message_bus_query_fetcher.py
@@ -3,7 +3,7 @@ from typing import Any, Self
 
 import pytest
 
-from forging_blocks.foundation.messages import MessageMetadata, Query
+from forging_blocks.domain.messages import MessageMetadata, Query
 from forging_blocks.infrastructure import MessageBusQueryFetcher
 from tests.fixtures.fake_message_bus import FakeMessageBus
